sp25_taxi_snowflake/requirements.txt
pandas
pyarrow<19.0.0
python-dotenv
snowflake-connector-python
